Refactor Ollama utilities
